[UI] - Fixed various graphical glitches on rocm device menu. - Fixed issue with device selection and re-entry to rocm device menu. - Fixed issue with uninstall and detection of package manager installations.

+++ runfile-installer/UI/src/menu_data.h
@@ -126,7 +126,7 @@ typedef struct _MENU_DATA
     // items
     int         curItemListIndex;
     ITEM_DATA   itemList[MAX_NUM_ITEM_LIST];
-    uint32_t    itemSelections;
+    uint64_t    itemSelections;
     int         curItemSelection;
     int         startListIndex;
     int         endListIndex;
